Refactored AuditTrail import/export.
