+++ src/Orchard.Web/Modules/Orchard.AuditTrail/Orchard.AuditTrail.csproj
@@ -194,6 +194,8 @@
     <Compile Include="Menus\RecycleBinAdminMenu.cs" />
     <Compile Include="Providers\AuditTrail\AuditTrailEventSettingEventData.cs" />
     <Compile Include="Providers\AuditTrail\AuditTrailManagerExtensions.cs" />
+    <Compile Include="Recipes\Builders\AuditTrailStep.cs" />
+    <Compile Include="Recipes\Executors\AuditTrailStep.cs" />
     <Compile Include="ViewModels\RecycleBinCommand.cs" />
     <Compile Include="Controllers\RecycleBinController.cs" />
     <Compile Include="Controllers\ContentController.cs" />
@@ -212,9 +214,6 @@
     <Compile Include="Helpers\HtmlExtensions.cs" />
     <Compile Include="Helpers\XmlHelper.cs" />
     <Compile Include="Helpers\StringExtensions.cs" />
-    <Compile Include="ImportExport\AuditTrailExportStep.cs" />
-    <Compile Include="ImportExport\AuditTrailExportHandler.cs" />
-    <Compile Include="ImportExport\AuditTrailImportHandler.cs" />
     <Compile Include="Models\AuditTrailEventRecordResult.cs" />
     <Compile Include="Models\AuditTrailTrimmingSettingsPart.cs" />
     <Compile Include="Providers\AuditTrail\AuditTrailTrimmingSettingsEventProvider.cs" />
@@ -329,6 +328,7 @@
   <ItemGroup>
     <Content Include="Views\AuditTrailEvent-Content-Restored.SummaryAdmin.cshtml" />
   </ItemGroup>
+  <ItemGroup />
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
     <VSToolsPath Condition="'$(VSToolsPath)' == ''">$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)</VSToolsPath>
